+++ examples/typescript/src/client.ts
@@ -87,15 +87,24 @@ async function subscribeCommand(client, args) {
 
   // Handle updates
   stream.on("data", (data) => {
-    if (data.transaction && args.transactionsParsed) {
+    if (
+      data.transaction &&
+      (args.transactionsParsed || args.transactionsDecodeErr)
+    ) {
       const slot = data.transaction.slot;
       const message = data.transaction.transaction;
-      const tx = txEncode.encode(message, txEncode.encoding.Json, 255, true);
-      console.log(
-        `TX filters: ${data.filters}, slot#${slot}, tx: ${JSON.stringify(tx)}`
-      );
-      const err = txErrDecode.decode(data.transaction.transaction.meta.err.err);
-      //   console.log(`TX error: ${inspect(err)}`);
+      if (args.transactionsParsed) {
+        const tx = txEncode.encode(message, txEncode.encoding.Json, 255, true);
+        console.log(
+          `TX filters: ${data.filters}, slot#${slot}, tx: ${JSON.stringify(tx)}`
+        );
+      }
+      if (message.meta.err && args.transactionsDecodeErr) {
+        const err = txErrDecode.decode(message.meta.err.err);
+        console.log(
+          `TX filters: ${data.filters}, slot#${slot}, err: ${inspect(err)}}`
+        );
+      }
       return;
     }
 
@@ -393,6 +402,11 @@ function parseCommandLineArgs() {
         "transactions-parsed": {
           default: false,
           describe: "parse transaction to json",
+          type: "boolean",
+        },
+        "transactions-decode-err": {
+          default: false,
+          describe: "decode transactions errors",
           type: "boolean",
         },
         "transactions-status": {