+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The minor version will be incremented upon a breaking change and the patch versi
 ### Features
 
 - proto: add tonic feature ([#474](https://github.com/rpcpool/yellowstone-grpc/pull/474))
+- nodejs: add parse err function ([#483](https://github.com/rpcpool/yellowstone-grpc/pull/483))
 
 ### Breaking